# openapi.yaml
openapi: 3.0.0
info:
  version: "1.0"
  title: "UniLinks Web API"
  description: "Web Resources Specification (A7) for UniLinks"
servers:
  - url: https://lbaw.fe.up.pt
    description: Production server
security: []
...
components:
  securitySchemes:
    cookieuser:
      type: apiKey
      in: cookie
      name: SESSIONID #cookie used to save session id
paths:
...
  /login:
    get:
      operationId: R101
      summary: "R101: Login Form"
      description: "Provide login form. Access: PUB"
      tags:
        - "M01"
      responses:
        "200":
          description: "Ok. Show log-in UI"
    post:
      operationId: R102
      summary: "R102: Login Action"
      description: "Processeses login credentials (form input) Access: PUB"
      tags:
        - "M01"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                _token:
                  type: string
              required:
                - email
                - password
                - _token
      responses:
        "302":
          description: "Redirect after processing the login credentials."
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: "Successful authentication. Redirect to user profile."
                  value: "/feed"
                302Error:
                  description: "Failed authentication. Redirect to login form."
                  value: "/login"
            Set-cookie:
              schema:
                type: string
                description: sessionID
                example: SESSIONID=abcde12345; Secure; Path=/; HttpOnly; SameSite=Strict
...
  /timeline:
    get:
      security:
        - cookieuser: []
      operationId: R302
      summary: "R302: View Feed"
      description: "Show the current logged user's feed. Access: Authenticated-USR, ADM"
      tags:
        - "M03"
      responses:
        "200":
          description: "Ok. Show feed"
...